Removal of platform-specific code from Core project, due to an internal helper class that must be unit-tested.
